refactor(Cars, Game) : Car 리스트를 가지고 있는 클래스 생성 -> Game 클래스 역할 분리
