+++ src/test/java/racingcar/ApplicationTest.java
@@ -1,6 +1,7 @@
 package racingcar;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
+
 import org.junit.jupiter.api.Test;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
@@ -9,33 +10,34 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ApplicationTest extends NsTest {
-    private static final int MOVING_FORWARD = 4;
-    private static final int STOP = 3;
+	private static final int MOVING_FORWARD = 4;
+	private static final int STOP = 3;
 
-    private static final String ERROR_MESSAGE = "[ERROR]";
+	private static final String ERROR_MESSAGE = "[ERROR]";
 
-    @Test
-    void 전진_정지() {
-        assertRandomNumberInRangeTest(
-            () -> {
-                run("pobi,woni", "1");
-                assertThat(output()).contains("pobi : -", "woni : ", "최종 우승자 : pobi");
-            },
-            MOVING_FORWARD, STOP
-        );
-    }
+	@Test
+	void 전진_정지() {
+		assertRandomNumberInRangeTest(
+			() -> {
+				run("pobi,woni", "1");
+				assertThat(output()).contains("pobi : -", "woni : ", "최종 우승자 : pobi");
+			},
+			MOVING_FORWARD, STOP
+		);
+	}
 
-    @Test
-    void 이름에_대한_예외_처리() {
-        assertSimpleTest(() ->
-            assertThatThrownBy(() -> runException("pobi,javaji"))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining(ERROR_MESSAGE)
-        );
-    }
+	@Test
+	void 이름에_대한_예외_처리() {
+		assertSimpleTest(
+			() -> {
+				runException("pobi,javaji");
+				assertThat(output()).contains(ERROR_MESSAGE);
+			}
+		);
+	}
 
-    @Override
-    public void runMain() {
-        Application.main(new String[]{});
-    }
+	@Override
+	public void runMain() {
+		Application.main(new String[] {});
+	}
 }
